+++ src/ts/src/public/account/AccountApiLoginRequest.ts
@@ -12,4 +12,8 @@ export interface AccountApiLoginRequest {
 
 export interface AccountApiLoginResponse extends WithToken {
   accountId: string;
+  /**
+   * When true the server wants you to change the password as soon as possible. This usually happens when you login with an temporary password.
+   */
+  askPasswordChange: boolean;
 }